+++ packages/paraglide/paraglide-js-adapter-sveltekit/example/svelte.config.js
@@ -1,23 +1,28 @@
 import adapter from "@sveltejs/adapter-static"
-import { vitePreprocess } from "@sveltejs/kit/vite"
+import { vitePreprocess } from "@sveltejs/vite-plugin-svelte"
+import { mdsvex } from "mdsvex"
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
-	preprocess: vitePreprocess(),
+	preprocess: [
+		vitePreprocess(),
+		mdsvex({
+			extension: ".svx",
+		}),
+	],
 	kit: {
 		adapter: adapter(),
 
 		alias: {
 			$paraglide: "./src/paraglide/",
+			$lib: "./src/lib/",
 		},
-
-		// Need for crawling to work until
-		// https://github.com/sveltejs/kit/issues/11133
-		// is fixed
-		prerender: {
-			entries: ["/"],
+		paths: {
+			base: "/base",
 		},
 	},
+
+	extensions: [".svelte", ".svx"],
 }
 
 export default config